Implement mutually exclusive bit
If we are turning on a gpio pin that is mutually exclusive,
go through all the other pins that are also mutually
exclusive and turn them off first.  This helps in situations
where you have low speed and high speed for your motor
connected to different gpio pins and you want to make sure
you can't accidentally turn them both on at the same time.