Fixing interval bugs.
Hopefully its done this time.  Not sure if its a bug as much as
very poor implementation.  Well, now the interval stuff seems to
be working (will see tomorrow for sure), while at the same time
manually turning things on will work regardless of the interval.
This is accomplished by having the cron job run the command with
the first action parameter being "cron-write", then checking
specifically for that when determining if we should run the job
or not based on the date.