Fix bug: never reset cron-write action to just write.
Maybe this time will be it.  Noticed things still werent working properly.
I was only reseting cron-write back to the write action in the cases where
we were not going to run the write action.  Well that didnt help much. Now
setting instances of cron-write back to write if the device pin is found.