Add feature to bypass interval if wind is 10mph or more.
Add the ability to optionally run a device if the wind is 10mph or more
even if it is not the day interval for the specific device.  For instance
have the sweep only run every other day unless there are 10mpg or more
winds, in which case run the sweep.